fieldworg, snippets: include proper replac cells if pol was missing *
Because of this, we now also support and return the true 'missing'
replac cells case.  Before, all cells from the level3 cell were
returned in this case, since true missingness is not distinguished from
'missing because polygon_id was missing'. Since the latter is no more
present (nor allowed), missingness is true missingness and returned as
such.